move CodebaseStatusProvider to separate file
